requirements.txt
Django==4.0.1
djangorestframework 
django-cors-headers
opencv-python